+++ packages/primeng/src/index.ts
@@ -4,3 +4,4 @@ export * from './lib/ngx-json-form-debug.component';
 export * from './lib/signature-pad.component';
 export * from './lib/time-slots.component';
 export * from './lib/markdown-editor.component';
+export * from './lib/captcha.component';