+++ packages/primeng/src/lib/ngx-form-engine-primeng.component.ts
@@ -43,6 +43,7 @@ import { TreeSelectModule } from 'primeng/treeselect';
 import { SignaturePadComponent } from './signature-pad.component';
 import { TimeSlotsComponent } from './time-slots.component';
 import { MarkdownEditorComponent } from './markdown-editor.component';
+import { CaptchaComponent } from './captcha.component';
 
 import {
   FieldRegistry,
@@ -95,6 +96,7 @@ import {
     SignaturePadComponent,
     TimeSlotsComponent,
     MarkdownEditorComponent,
+    CaptchaComponent,
   ],
 })
 export class NgxJsonFormComponent {